+++ packages/nextjs-config/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @posthog/nextjs-config
 
+## 1.9.20
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @posthog/webpack-plugin@1.4.20
+
 ## 1.9.19
 
 ### Patch Changes