+++ packages/rrweb/all/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @rrweb/all
 
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies []:
+    - @posthog/rrweb@0.0.63
+
 ## 0.0.1
 
 ### Patch Changes